requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
playwright==1.40.0
pydantic==2.5.0
openai==1.3.0
pandas==2.1.3
PyPDF2==3.0.1
